requirements.txt
google-genai
pypdf
nltk
tiktoken
chromadb
python-dotenv